requirements.txt
#!/usr/bin/env python3
# -*- coding: utf-8 -*-


tweepy>=4.12.0
openai>=1.0.0
pillow>=9.0.0
matplotlib>=3.5.0
requests>=2.27.0
schedule>=1.1.0
PyQt5>=5.15.0
python-dotenv>=0.19.0
pandas>=1.3.0
